Update Movement, City placement, border control
I was able to update the movement to be a little more complex. it now has a way to tell with the player is holding down the keys, and when not. Slows when not, boulders when opposite velocity direction input is placed. And updated the placement of the cities to be spread out more.